random_forest/pyproject.toml
[tool.poetry]
name = "micro-random-forest-tuning"
version = "0.1.0"
description = ""
authors = ["Soumendra Dhanee <soumendra@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
streamlit = "^0.73.1"
numpy = "^1.19.4"
pandas = "^1.1.5"
scikit-learn = "^0.24.0"
xgboost = "^1.3.1"
lightgbm = "^3.1.1"
catboost = "^0.24.3"
attrs = "^20.3.0"
plotly = "^4.14.1"
ucimlr = "^0.3.0"
openml = "^0.11.0"
hiplot = "^0.1.22"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.2.1"
mypy = "^0.790"
flake8 = "^3.8.4"
pylint = "^2.6.0"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
